Fix 5 real bugs found by reading the test instance's own IIS logs
Root-caused and fixed directly from stdout_*.log on the machine hosting
the NormantonNexus test site, now that a dotnet SDK + build/test are
available here for the first time this session:

- Missed Reversals page ("scrap-reversal/missed"/"search"): the
  ScrapDocSql BackflushReversed column was CASE WHEN...THEN 1 ELSE 0 END
  with no CAST, which SQL Server types as INT - Dapper's strict record
  binding needs an exact match against bool. This was the "missed
  reversals page has an error" report from earlier today with no
  stack trace available at the time. Same fix (CAST ... AS BIT) applied
  to two more unreported instances of the identical pattern found by
  grepping for it: DeliveryCompletionHelper.IsOwned (ZDELFLAG pallet
  query) and DbExplorerHelper.IsPrimaryKey.

- ProductionScheduleHelper.DiffProductionScheduleOtifAsync: .ToDictionary()
  threw on a duplicate (ReferenceDocument, Item) key where Node's real
  `new Map(...)` silently keeps the last entry. Replaced with a manual
  last-wins loop matching Node's actual semantics - most likely a
  downstream symptom of the RunFullRefreshAsync race fixed earlier today.

- MaterialRequestUnitsHelper.ListAllAsync selected 7 columns but
  MaterialRequestUnitRow only declared 4 (the shape GetByMaterialAsync
  genuinely needs, matching Node's own narrower route). Added a second
  record, MaterialRequestUnitAdminRow, for the admin-list shape instead
  of widening the shared one.

- AuditLogRow.LogId was declared long; the real PortalAuditLog.LogID
  column is a plain INT identity. Narrowed the record's type.

700 tests, dotnet build/test both clean.

+++ dotnet/NormantonNexus/Models/Dto/UserAdminModels.cs
@@ -50,7 +50,7 @@ public sealed record BulkStatusResult(IReadOnlyList<BulkStatusRowResult> Results
 
 public sealed record AuditLogQuery(string? Event, string? Username, string? Detail, DateTime? From, DateTime? To);
 
-public sealed record AuditLogRow(long LogId, DateTime EventTime, string? Username, string EventType, string? Detail, string? IpAddress);
+public sealed record AuditLogRow(int LogId, DateTime EventTime, string? Username, string EventType, string? Detail, string? IpAddress);
 
 public sealed record PermissionDefinitionRow(string PermissionCode, string PermissionName, string? Description, string Category, DateTime CreatedAt);
